Fix: gzipped SVGs were not marked correctly.
Ember-Deploy gzips SVGs by default and this should be
reflected in the s3-adapter. S3 will now take the
`config.assets.gzipExtension`-property to determine which
files will be marked as gzipped on upload.

If no gzipExtension property was defined in `deploy.json`
ember-deploy-s3 will fallback to the ember-deploy defaults
('js', 'css', 'svg').